+++ src/main.rs
@@ -81,9 +81,8 @@ fn main() -> Result<()> {
     }
     let args: Args = argh::cargo_from_env();
     let source_tree = SourceTree::new(&args.dir)?;
-    let console = console::Console::new()
-        .show_all_logs(args.all_logs)
-        .show_times(!args.no_times);
+    let mut console = console::Console::new().show_all_logs(args.all_logs);
+    console.set_show_times(!args.no_times);
     let options = Options::from(&args);
     interrupt::install_handler();
     if args.list {