+++ src/options.rs
@@ -7,7 +7,7 @@ use std::time::Duration;
 use crate::*;
 
 /// Options for running experiments.
-#[derive(Default, Debug)]
+#[derive(Default, Debug, Clone)]
 pub struct Options {
     /// Don't run the tests, just see if each mutant builds.
     pub check_only: bool,
@@ -17,10 +17,18 @@ pub struct Options {
 
 impl Options {
     /// Return the maximum run time for `cargo test` commands.
-    ///
+    ///commands
     /// Build and check are not affected.
     pub fn test_timeout(&self) -> Duration {
         self.test_timeout
+    }
+
+    pub fn has_test_timeout(&self) -> bool {
+        self.test_timeout < Duration::MAX
+    }
+
+    pub fn set_test_timeout(&mut self, test_timeout: Duration) {
+        self.test_timeout = test_timeout;
     }
 }