+++ src/outcome.rs
@@ -146,4 +146,14 @@ impl Outcome {
             && self.last_phase() == Phase::Test
             && self.last_phase_result().success()
     }
+
+    /// Duration of the test phase, if tests were run.
+    pub fn test_duration(&self) -> Option<Duration> {
+        if let Some(phase_result) = self.phase_results.last() {
+            if phase_result.phase == Phase::Test {
+                return Some(phase_result.duration);
+            }
+        }
+        None
+    }
 }