+++ NEWS.md
@@ -4,9 +4,14 @@
 
 - A new `--timeout SECS` option to limit the runtime of any `cargo test`
   invocation, so that mutations that cause tests to hang don't cause
-  `cargo mutants` to hang. There is no timeout by default. On Unix, the `cargo`
-  subprocesses run in a new process group. As a consequence ctrl-c is explicitly
-  caught and propagated to the child processes.
+  `cargo mutants` to hang. 
+
+  A default timeout is set based on the time to run tests in an unmutated tree.
+  There is no timeout by default on the unmutated tree. 
+
+  On Unix, the `cargo` subprocesses run in a new process group. As a
+  consequence ctrl-c is explicitly caught and propagated to the child
+  processes.
 
 - Show a progress bar while looking for mutation opportunities, and show the
   total number found.